# services/billing-service/openapi.yaml
openapi: 3.1.0
info:
  title: FuzeFront Billing Service API
  version: 1.3.0
  description: >-
    REST API for the FuzeFront billing-service. Bridges platform entities
    (users / organizations, referenced by `(entityType, entityId)` only) to
    payment-provider customers, subscriptions, payment-method setup, usage metering, and
    manual credits.

    This contract is **derived from the real route handlers** in
    `services/billing-service/src/routes/*` and `src/app.ts` and is the single
    source of truth for `@fuzefront/billing-client`. Changing the API means
    amending this file first (see the `api-contract-first` skill).

    Card data never flows through this API — only provider ids and client secrets.

    ### Async surface (events)
    Beyond HTTP, the service emits Kafka events whose Zod schemas live in
    `shared/src/kafka/schemas/`:
      * `billing.subscription.changed` — `billingSubscriptionChangedSchemaV1`
      * `billing.usage.recorded`       — `billingUsageRecordedSchemaV1`
      * `billing.payment.completed`    — `billingPaymentCompletedSchemaV1`
        (one-time payment-mode checkout outcome: paid / failed / expired)
      * `billing.trial.ending`  / `billing.payment.failed` (raw notify envelopes)
    Consumers (e.g. the backend plan-state projection) read
    `billing.subscription.changed` rather than reading the billing DB directly.
    Consumer products fulfilling one-time orders (e.g. `mendys-datasets`) read
...
  license:
    name: UNLICENSED
servers:
...
paths:
...
  /payments/checkout:
    post:
      operationId: createPaymentCheckoutSession
      tags: [payments]
      summary: Create a hosted Checkout Session (one-time payment mode)
      description: >-
        Creates a hosted Checkout Session in `payment` mode for an
        allowlisted consumer product (e.g. `mendys-datasets`) and returns the
        provider-hosted `url` to redirect the buyer to. The caller supplies the
        priced line items (name + unit amount in cents + quantity); the service
        builds provider `price_data` from them — no provider price catalogue entry
        is required. On completion the provider fires `checkout.session.completed`
        (mode `payment`), which upserts the local `billing.payments` mirror and
        emits `billing.payment.completed` on Kafka. Failure/expiry are mirrored
        via `payment_intent.payment_failed` / `checkout.session.expired`.
        Mirrors `POST /payments/checkout`.


        Authorization (money path): identical to `/checkout` — the host proxy
        authenticates the platform JWT, authorizes the caller against the
        target entity, and forwards the SERVER-DERIVED actor + entity as
        trusted headers (`X-Billing-Actor-User-Id`, `X-Billing-Entity-Type`,
        `X-Billing-Entity-Id`; aliases `X-FF-Actor-Id` / `X-FF-Org-Id`). The
        service RE-VERIFIES that the body's `entityType`/`entityId` match the
        authorized entity before creating any provider object (mismatch → 403).
        `productKey` must be in the server-side allowlist
        (`BILLING_PRODUCT_KEYS`), `currency` in the configured allowlist
        (default `usd`, `eur`), and each line amount plus the order total are
        bounded server-side (`BILLING_PAYMENT_MAX_TOTAL_CENTS`, default
        5,000,000 cents) — violations → 400.


        Idempotency: retries with IDENTICAL parameters dedupe to one provider
        session (param-fingerprint idempotency key, same scheme as
        `/checkout`); any changed parameter yields a fresh session.
      security:
        - internalToken: []
      parameters:
        - $ref: '#/components/parameters/ActorUserIdHeader'
        - $ref: '#/components/parameters/EntityTypeHeader'
        - $ref: '#/components/parameters/EntityIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCheckoutRequest'
      responses:
        '200':
          description: Payment-mode Checkout Session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentCheckoutResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '422':
          $ref: '#/components/responses/IdentityError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Idempotency-key conflict (same key, different params).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderErrorBody'
        '502':
          $ref: '#/components/responses/ProviderError'
  /payments/sessions/{sessionId}:
    parameters:
      - name: sessionId
        in: path
        required: true
        description: The hosted Checkout Session id (`cs_...`).
        schema:
          type: string
    get:
      operationId: getPaymentSession
      tags: [payments]
      summary: Get the local payment mirror for a Checkout Session
      x-pagination: exempt
      description: >-
        Returns the `billing.payments` mirror row for a payment-mode Checkout
        Session — the reconciliation-polling surface for consumer services
        that missed (or cannot receive) the `billing.payment.completed` event.
        `status` is `pending` until a webhook lands, then `paid` / `failed` /
        `expired`. 404 when the session is unknown. When the local row is
        missing but the session exists at the provider AND its `productKey` metadata
        is allowlisted, the service falls back to the live provider, re-mirrors the
        row, and returns it (singleton resource — pagination exempt).


        Authorization: same model as `/payments/checkout`; the proxy-authorized
        entity must OWN the session (mismatch → 403).
      security:
        - internalToken: []
      parameters:
        - $ref: '#/components/parameters/ActorUserIdHeader'
        - $ref: '#/components/parameters/EntityTypeHeader'
        - $ref: '#/components/parameters/EntityIdHeader'
      responses:
        '200':
          description: The payment mirror row.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [payment]
                properties:
                  payment:
                    $ref: '#/components/schemas/BillingPayment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/ProviderError'
...
components:
  securitySchemes:
    internalToken:
      type: http
      scheme: bearer
      description: >-
        `BILLING_INTERNAL_TOKEN` presented as a Bearer token by the backend /
        `@fuzefront/billing-client` on all non-public routes. Verified by
        `requireInternalToken`.
  parameters:
...
    ActorUserIdHeader:
      name: X-Billing-Actor-User-Id
      in: header
      required: true
      description: >-
        Authenticated actor (the platform user id) — set by the host proxy
        AFTER it authenticates + authorizes the caller. Short alias:
        `X-FF-Actor-Id`.
      schema:
        type: string
    EntityTypeHeader:
      name: X-Billing-Entity-Type
      in: header
      required: true
      description: >-
        Server-derived authorized entity type (`user`|`organization`) set by the
        proxy. When the `X-FF-Org-Id` alias is used this is implied to be
        `organization`.
      schema:
        $ref: '#/components/schemas/EntityType'
    EntityIdHeader:
      name: X-Billing-Entity-Id
      in: header
      required: true
      description: >-
        Server-derived authorized entity id set by the proxy. The service
        re-verifies the request target matches this. Short alias: `X-FF-Org-Id`.
      schema:
        type: string
        format: uuid
  responses:
    IdentityError:
      description: >-
        The create body violates the server-owned identifier standard — it
        supplied an `id`, or its `lid` graph is malformed or reaches outside
        billing's aggregate. Returned by the graph-create middleware mounted in
        `src/app.ts`, before the route handler runs.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GraphCreateError'
    Unauthorized:
      description: Missing or invalid internal token, or missing actor context.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: >-
        The authenticated/authorized caller may not act on the target object
        (e.g. organizationId does not match the authorized entity, or the
        operation requires platform-admin context).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Request body failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorBody'
    ProviderError:
      description: Upstream provider error (bad gateway).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProviderErrorBody'
...
  schemas:
    EntityType:
      type: string
      enum: [user, organization]
      description: Which platform entity owns the billing relationship.
...
    GraphCreateError:
      type: object
      additionalProperties: false
      required: [error, code, message, path]
      description: 422 returned when a create graph violates the identifier standard.
      properties:
        error:
          type: string
          enum: [unprocessable_entity]
        code:
          type: string
          enum:
            - CLIENT_SUPPLIED_ID
            - UNKNOWN_LID
            - DUPLICATE_LID
            - MALFORMED_LID
            - MISSING_TYPE
            - UNKNOWN_TYPE
            - CROSS_AGGREGATE_LID
            - GRAPH_TOO_LARGE
            - GRAPH_TOO_DEEP
        message:
          type: string
        path:
          type: string
          description: Location of the offending node; empty string is the document root.
          examples: ['/invoices/0']
...
    PaymentStatus:
      type: string
      enum: [pending, paid, failed, expired]
      description: >-
        Lifecycle of a payment-mode Checkout Session mirror: `pending` from
        session creation until a webhook lands; then `paid`
        (checkout.session.completed, paid), `failed`
        (payment_intent.payment_failed) or `expired`
        (checkout.session.expired). `paid` is terminal and never downgraded.
    PaymentLineItem:
      type: object
      additionalProperties: false
      required: [name, unitAmountCents, quantity]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 250
          description: Line-item display name shown on the provider-hosted page.
        description:
          type: string
          minLength: 1
          maxLength: 500
          description: Optional line-item description shown under the name.
        unitAmountCents:
          type: integer
          minimum: 1
          description: >-
            Unit price in the currency's minor unit (cents). Bounded
            server-side together with the order total
            (`BILLING_PAYMENT_MAX_TOTAL_CENTS`).
        quantity:
          type: integer
          minimum: 1
          description: Units of this line item.
    PaymentCheckoutRequest:
      type: object
      additionalProperties: false
      required:
        - productKey
        - externalOrderId
        - entityType
        - entityId
        - currency
        - lineItems
        - successUrl
        - cancelUrl
      properties:
        productKey:
          type: string
          minLength: 1
          description: >-
            Consumer product key (e.g. `mendys-datasets`). MUST be in the
            server-side allowlist (`BILLING_PRODUCT_KEYS`); unknown keys → 400.
        externalOrderId:
          type: string
          minLength: 1
          maxLength: 200
          description: >-
            The consumer product's own order id. Stamped on the session as
            `client_reference_id` + metadata and echoed on
            `billing.payment.completed` so the product can correlate.
        entityType:
          $ref: '#/components/schemas/EntityType'
        entityId:
          type: string
          format: uuid
          description: >-
            Paying entity. MUST match the proxy-authorized entity (re-verified
            server-side; mismatch → 403).
        currency:
          type: string
          minLength: 3
          maxLength: 3
          description: >-
            ISO 4217 currency code (case-insensitive; lowercased server-side).
            Must be in the configured allowlist (default `usd`, `eur`).
          examples: [usd, eur]
        lineItems:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/PaymentLineItem'
        successUrl:
          type: string
          format: uri
          description: Where the provider redirects after a completed payment.
        cancelUrl:
          type: string
          format: uri
          description: Where the provider redirects if the buyer cancels.
    PaymentCheckoutResponse:
      type: object
      additionalProperties: false
      required: [sessionId, url]
      properties:
        sessionId:
          type: string
          description: The hosted Checkout Session id (`cs_...`).
        url:
          type: [string, 'null']
          description: provider-hosted Checkout URL to redirect the buyer to.
    BillingPayment:
      type: object
      additionalProperties: false
      required:
        - id
        - sessionId
        - paymentIntentId
        - productKey
        - externalOrderId
        - entityType
        - entityId
        - amountTotalCents
        - currency
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Local billing.payments primary key.
        sessionId:
          type: string
          description: hosted Checkout Session id (`cs_...`).
        paymentIntentId:
          type: [string, 'null']
          description: >-
            provider PaymentIntent id (`pi_...`); null until the provider creates the
            intent (typically when the buyer reaches the payment page).
        productKey:
          type: string
        externalOrderId:
          type: string
        entityType:
          $ref: '#/components/schemas/EntityType'
        entityId:
          type: string
          format: uuid
        amountTotalCents:
          type: integer
          minimum: 0
          description: Order total in the currency's minor unit (cents).
        currency:
          type: string
          description: ISO 4217 currency code, lowercased.
          examples: [usd]
        status:
          $ref: '#/components/schemas/PaymentStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
    ProviderErrorBody:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
          examples: ['provider error']
        message:
          type: string
    ValidationErrorBody:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
          examples: ['invalid request']
        details:
          description: Zod flatten() output describing the field errors.